refactor plots and tables
